move visual properties
